requirements.txt
pandas==2.0.2
mysql-connector==2.2.9
requests==2.31.0
numpy==1.26.3
gspread==5.9.0
google.auth==2.20.0
pydrive==1.3.1
gspread_dataframe==3.3.1
python-decouple==3.8
woocommerce
